move gui into a separate directory
