+++ otsdaq-mu2e-tracker/CMakeLists.txt
@@ -2,3 +2,4 @@ add_subdirectory(FEInterfaces)
 add_subdirectory(Generators)
 add_subdirectory(ArtModules)
 add_subdirectory(ui)
+add_subdirectory(gui)
