+++ otsdaq-mu2e-tracker/gui/root/classes.h
@@ -1,0 +1,4 @@
+#include <memory> // std::shared_ptr, std::make_shared
+#include <cstring> // memcpy
+
+#include "otsdaq-mu2e-tracker/gui/DtcGui.hh"
