+++ otsdaq-mu2e-tracker/gui/root/classes_def.xml
@@ -1,0 +1,5 @@
+<lcgdict>
+
+  <class name="DtcGui" />
+
+</lcgdict>
