+++ otsdaq-mu2e-tracker/ui/DtcInterface.hh
@@ -35,15 +35,16 @@ namespace trkdaq {
     DTCLib::DTC* Dtc() { return fDtc; }
 
                                         // EWLength - in 25 ns ticks
-    void         InitEmulatedCFOMode(int EWLength, int NMarkers, int FirstEWTag);
-    void         InitExternalCFOMode();
+    void         InitEmulatedCFOReadoutMode(int EWLength, int NMarkers, int FirstEWTag);
+    void         InitExternalCFOReadoutMode();
                                         // launch "run plan" in emulated mode...
     void         LaunchRunPlan (int NEvents );
 
-    void         PrintBuffer   (const void* ptr, int nw);
-    void         PrintRegister (uint16_t Register, const char* Title = "");
-    void         PrintRocStatus(int Link);
-    void         PrintStatus   ();
+    void         PrintBuffer     (const void* ptr, int nw);
+    void         PrintFireflyTemp();
+    void         PrintRegister   (uint16_t Register, const char* Title = "");
+    void         PrintRocStatus  (int Link);
+    void         PrintStatus     ();
 
     uint32_t     ReadRegister    (uint16_t Register);
     void         ReadSubevents   (std::vector<std::unique_ptr<DTCLib::DTC_SubEvent>>& Vsev,
